dra e2e: CRDs for class and claim parameters
The driver now supports a ConfigMap (as before) and CRDs (new) for class and
claim parameters objects. The CRDs are defined so that they are compatible with
the "counter" numeric parameter model.

The CRDs were generated from the api/v1alpha1/types.go file with

    controller-gen crd paths=./test/e2e/dra/test-driver/api/... output:dir=test/e2e/dra/test-driver/deploy/crd

There's no update+verify script for these CRDs because that would imply that
k/k has to depend on controller-gen, which itself depends on staging repos. To
minimize the amount of generated code, the tests and test driver use a dynamic
client to handle these CRDs.

Code which needs access to these CRDs can do so via the embed.FS in the deploy package.

+++ test/integration/scheduler_perf/dra_test.go
@@ -194,7 +194,7 @@ func (op *createResourceDriverOp) run(ctx context.Context, tb testing.TB, client
 	resources := draapp.Resources{
 		DriverName:     op.DriverName,
 		NodeLocal:      true,
-		MaxAllocations: op.MaxClaimsPerNode,
+		MaxAllocations: int64(op.MaxClaimsPerNode),
 	}
 
 	nodes, err := client.CoreV1().Nodes().List(ctx, metav1.ListOptions{})
